Added a unit test for penColor and an example of setting the robot to a local image file.
